requirements.txt
requests
PyYAML
Flask
Flask-SocketIO
eventlet
beautifulsoup4
arxiv